requirements.txt
flask
webdavclient3
Pillow
redis
requests
